Add files via upload
Classroom Grading

Firstly we will ask the user the number system in which they are going to give inputs. A classroom has n(user input) students. It is required to rearrange them in descending order according to their grades in the course and also grades the students (A for 85+, B for 75-85 and so on).

The inputs are two columns. The first one will get input of the student IDs and the second one will get input of their Marks.

All the output values will be in decimal and grades will be like A,B,C,D,F.

The outputs are three columns. The first one contains the student number arranged according to their marks, the second column shows these marks and the third column shows the grades according to the marks.